+++ PropertyTaxStrategy.h
@@ -6,7 +6,7 @@
 class PropertyTaxStrategy : public TaxStrategy
 {
     public:
-        double calculateTax(int total) override
+        int calculateTax(int total) override
         {
             return total * (2/100);            
         };
